+++ invenio_theme_iform/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2020-2022 Graz University of Technology.
+#
+# invenio-theme-iform is free software; you can redistribute it and/or
+# modify it under the terms of the MIT License; see LICENSE file for more
+# details.
+
+"""invenio module for I-Form theme."""
+
+from .ext import InvenioThemeIform
+
+__version__ = "4.13.0"
+
+__all__ = ("__version__", "InvenioThemeIform")